requirements.txt
anyio==3.6.1
boto3==1.24.43
botocore==1.27.43
click==8.1.3
fastapi==0.79.0
h11==0.13.0
idna==3.3
imageio==2.21.0
jmespath==1.0.1
networkx==2.8.5
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pydantic==1.9.1
pyparsing==3.0.9
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.20.0
PyWavelets==1.3.0
s3transfer==0.6.0
scikit-image==0.19.3
scipy==1.9.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tifffile==2022.7.31
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
